# Speed up WithAnnotationsTracer x5.4 times in case of Dictionary arg.