--- a/Vostok.Tracing.Abstractions.Tests/Vostok.Tracing.Abstractions.Tests.csproj
+++ b/Vostok.Tracing.Abstractions.Tests/Vostok.Tracing.Abstractions.Tests.csproj
@@ -9,6 +9,7 @@
   <ItemGroup>
     <PackageReference Include="FluentAssertions" Version="5.4.1" />
     <PackageReference Include="NSubstitute" Version="3.1.0" />
+    <PackageReference Include="BenchmarkDotNet" Version="0.13.1" />
   </ItemGroup>
 
   <ItemGroup>