In case of ```WithAnnotationTracerExtensions.WithAnnotations(IReadOnlyDictionary<string, object> annotations...)``` extension, we call LINQ methods to transform ReadonlyDictionary to IEnumerable every time we call `Log` in the WithAnnotationsTracer class.

Fix this shameful decision by caching this transformation. This could be safe because dictionary is readonly.
And it's not too bad if someone calls WithAnnotations extension very often.


BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19043.1415 (21H1/May2021Update)
Intel Core i7-4771 CPU 3.50GHz (Haswell), 1 CPU, 8 logical and 4 physical cores
.NET SDK=6.0.101
  [Host]     : .NET 6.0.1 (6.0.121.56705), X64 RyuJIT
  DefaultJob : .NET 6.0.1 (6.0.121.56705), X64 RyuJIT


|      Method |      Mean |    Error |   StdDev | Ratio |  Gen 0 | Allocated |
|------------ |----------:|---------:|---------:|------:|-------:|----------:|
| WithDictOld | 207.70 ns | 4.088 ns | 3.824 ns |  1.00 | 0.0286 |     120 B |
| WithDictNew |  38.64 ns | 0.756 ns | 1.853 ns |  0.19 | 0.0076 |      32 B |
|    WithFunc |  38.06 ns | 0.799 ns | 1.500 ns |  0.18 | 0.0076 |      32 B |